+++ planning/autoware_velocity_smoother/src/smoother/analytical_jerk_constrained_smoother/velocity_planning_utils.cpp
@@ -217,7 +217,7 @@ bool calcStopVelocityWithConstantJerkAccLimit(
 
   if (xs.empty()) {
     for (size_t i = start_index; i < output_trajectory.size(); ++i) {
-      output_trajectory.at(i).longitudinal_velocity_mps = decel_target_vel;
+      output_trajectory.at(i).longitudinal_velocity_mps = static_cast<float>(decel_target_vel);
       output_trajectory.at(i).acceleration_mps2 = 0.0;
     }
     return true;
@@ -254,11 +254,12 @@ bool calcStopVelocityWithConstantJerkAccLimit(
   const auto jerk_at_wp = autoware::interpolation::lerp(xs, js, distances);
 
   for (size_t i = 0; i < vel_at_wp.size(); ++i) {
-    output_trajectory.at(start_index + i).longitudinal_velocity_mps = vel_at_wp.at(i);
-    output_trajectory.at(start_index + i).acceleration_mps2 = acc_at_wp.at(i);
+    output_trajectory.at(start_index + i).longitudinal_velocity_mps =
+      static_cast<float>(vel_at_wp.at(i));
+    output_trajectory.at(start_index + i).acceleration_mps2 = static_cast<float>(acc_at_wp.at(i));
   }
   for (size_t i = start_index + vel_at_wp.size(); i < output_trajectory.size(); ++i) {
-    output_trajectory.at(i).longitudinal_velocity_mps = decel_target_vel;
+    output_trajectory.at(i).longitudinal_velocity_mps = static_cast<float>(decel_target_vel);
     output_trajectory.at(i).acceleration_mps2 = 0.0;
   }
